prediction/tools/model_warmup/requirements.txt
requests
tensorflow-serving-api==1.15.0